first code

+++ BE-InD/src/main/java/fu/se/beind/Dto/request/ReviewRequest.java
@@ -1,8 +1,6 @@
 package fu.se.beind.Dto.request;
 
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.*;
 import lombok.Data;
 
 @Data
@@ -20,7 +18,7 @@ public class ReviewRequest {
     @NotNull(message = "Rating không được để trống")
     @Min(value = 1, message = "Rating tối thiểu là 1")
     @Max(value = 5, message = "Rating tối đa là 5")
-    private Integer rating;
+    private Byte rating;
 
     private String title;
